+++ tests/llmcompressor/transformers/finetune/test_oneshot_then_finetune.py
@@ -1,3 +1,4 @@
+import os
 import shutil
 import unittest
 from pathlib import Path
@@ -156,4 +157,5 @@ class TestOneshotThenFinetune(unittest.TestCase):
             )
 
     def tearDown(self):
-        shutil.rmtree(self.output)
+        if os.path.isdir(self.output):
+            shutil.rmtree(self.output)
